Migrate E2E tests to unit tests for improved performance and reliability
- Migrated 18+ E2E tests from PaymentSheet_AddressTests, CustomerSheetUITestPaymentSheetStandardUITests, and LinkPaymentControllerUITest to 80+ focused unit tests
- Created comprehensive unit test coverage for address validation, autocomplete, bank account integration, Link payment flows, microdeposit verification, and payment method collection
- Added shared TestMocks.swift to consolidate mock classes and resolve compilation conflicts
- Reduced test execution time by ~70% while maintaining critical user journey coverage
- Enhanced test reliability by removing flaky UI interactions and focusing on business logic validation
