+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/AddressAutocompleteTests.swift
@@ -12,233 +12,135 @@ import XCTest
 
 class AddressAutocompleteTests: XCTestCase {
 
-    // MARK: - Address Autocomplete Logic Tests
+    // MARK: - Address Autocomplete Functionality Tests
     
-    func testAddressAutocomplete_unitedStatesQuery() {
-        let mockAutocompleteService = MockAddressAutocompleteService()
-        let query = "354 Oyster Point"
+    func testAddressAutocomplete_supportedCountries() {
+        // Test that we can identify supported countries for autocomplete
+        let supportedCountries = ["US", "CA", "GB", "AU"]
         
-        mockAutocompleteService.searchAddresses(query: query, countryCode: "US") { results in
-            XCTAssertEqual(results.count, 1)
-            let result = results.first!
-            XCTAssertEqual(result.line1, "354 Oyster Point Blvd")
-            XCTAssertEqual(result.city, "South San Francisco")
-            XCTAssertEqual(result.state, "California")
-            XCTAssertEqual(result.postalCode, "94080")
-            XCTAssertEqual(result.country, "US")
+        for country in supportedCountries {
+            XCTAssertTrue(isAutocompleteSupported(countryCode: country), "Autocomplete should be supported for \(country)")
         }
     }
     
-    func testAddressAutocomplete_unsupportedCountry() {
-        let mockAutocompleteService = MockAddressAutocompleteService()
-        let query = "1 South Bay Parade"
+    func testAddressAutocomplete_unsupportedCountries() {
+        // Test that we correctly identify unsupported countries
+        let unsupportedCountries = ["NZ", "IN", "BR"]
         
-        mockAutocompleteService.searchAddresses(query: query, countryCode: "NZ") { results in
-            XCTAssertEqual(results.count, 0)
+        for country in unsupportedCountries {
+            XCTAssertFalse(isAutocompleteSupported(countryCode: country), "Autocomplete should not be supported for \(country)")
         }
     }
     
-    func testAddressAutocomplete_emptyQuery() {
-        let mockAutocompleteService = MockAddressAutocompleteService()
-        let query = ""
+    func testAddressSearch_validQuery() {
+        // Test that valid search queries can be processed
+        let validQueries = ["354 Oyster Point", "123 Main Street", "1600 Pennsylvania"]
         
-        mockAutocompleteService.searchAddresses(query: query, countryCode: "US") { results in
-            XCTAssertEqual(results.count, 0)
+        for query in validQueries {
+            XCTAssertTrue(isValidSearchQuery(query), "Query '\(query)' should be valid")
         }
     }
     
-    func testAddressAutocomplete_shortQuery() {
-        let mockAutocompleteService = MockAddressAutocompleteService()
-        let query = "35"
+    func testAddressSearch_invalidQuery() {
+        // Test that invalid search queries are rejected
+        let invalidQueries = ["", "1", "a", "  "]
         
-        mockAutocompleteService.searchAddresses(query: query, countryCode: "US") { results in
-            XCTAssertEqual(results.count, 0)
+        for query in invalidQueries {
+            XCTAssertFalse(isValidSearchQuery(query), "Query '\(query)' should be invalid")
         }
     }
     
-    // MARK: - Address Parsing Tests
-    
-    func testAddressParsing_completeUSAddress() {
-        let rawAddress = "354 Oyster Point Blvd, South San Francisco, CA 94080, USA"
-        let parsedAddress = AddressParser.parse(rawAddress: rawAddress)
+    func testAddressParsing_standardFormat() {
+        // Test parsing a standard US address format
+        let addressString = "354 Oyster Point Blvd, South San Francisco, CA 94080"
+        let components = parseAddressString(addressString)
         
-        XCTAssertEqual(parsedAddress.line1, "354 Oyster Point Blvd")
-        XCTAssertEqual(parsedAddress.city, "South San Francisco")
-        XCTAssertEqual(parsedAddress.state, "California")
-        XCTAssertEqual(parsedAddress.postalCode, "94080")
-        XCTAssertEqual(parsedAddress.country, "US")
+        XCTAssertEqual(components["street"], "354 Oyster Point Blvd")
+        XCTAssertEqual(components["city"], "South San Francisco")
+        XCTAssertEqual(components["state"], "CA")
+        XCTAssertEqual(components["postalCode"], "94080")
     }
     
-    func testAddressParsing_incompleteAddress() {
-        let rawAddress = "354 Oyster Point Blvd"
-        let parsedAddress = AddressParser.parse(rawAddress: rawAddress)
+    func testAddressParsing_internationalFormat() {
+        // Test parsing an international address format
+        let addressString = "10 Downing Street, London SW1A 2AA, United Kingdom"
+        let components = parseAddressString(addressString)
         
-        XCTAssertEqual(parsedAddress.line1, "354 Oyster Point Blvd")
-        XCTAssertNil(parsedAddress.city)
-        XCTAssertNil(parsedAddress.state)
-        XCTAssertNil(parsedAddress.postalCode)
-        XCTAssertNil(parsedAddress.country)
+        XCTAssertEqual(components["street"], "10 Downing Street")
+        XCTAssertEqual(components["city"], "London")
+        XCTAssertEqual(components["postalCode"], "SW1A 2AA")
     }
     
-    // MARK: - Address Autocomplete Availability Tests
-    
-    func testAddressAutocompleteAvailability_supportedCountries() {
-        XCTAssertTrue(AddressAutocompleteService.isSupported(countryCode: "US"))
-        XCTAssertTrue(AddressAutocompleteService.isSupported(countryCode: "CA"))
-        XCTAssertTrue(AddressAutocompleteService.isSupported(countryCode: "GB"))
-        XCTAssertTrue(AddressAutocompleteService.isSupported(countryCode: "AU"))
+    func testAddressValidation_completeAddress() {
+        // Test validation of a complete address from autocomplete
+        let address = PaymentSheet.Address(
+            city: "South San Francisco",
+            country: "US",
+            line1: "354 Oyster Point Blvd",
+            postalCode: "94080",
+            state: "California"
+        )
+        
+        XCTAssertNotNil(address.city)
+        XCTAssertNotNil(address.country)
+        XCTAssertNotNil(address.line1)
+        XCTAssertNotNil(address.postalCode)
+        XCTAssertNotNil(address.state)
     }
     
-    func testAddressAutocompleteAvailability_unsupportedCountries() {
-        XCTAssertFalse(AddressAutocompleteService.isSupported(countryCode: "NZ"))
-        XCTAssertFalse(AddressAutocompleteService.isSupported(countryCode: "IN"))
-        XCTAssertFalse(AddressAutocompleteService.isSupported(countryCode: "BR"))
+    func testAddressValidation_incompleteAddress() {
+        // Test validation of an incomplete address from autocomplete
+        let address = PaymentSheet.Address(
+            city: nil,
+            country: "US",
+            line1: "354 Oyster Point Blvd",
+            postalCode: nil,
+            state: nil
+        )
+        
+        XCTAssertNil(address.city)
+        XCTAssertNotNil(address.country)
+        XCTAssertNotNil(address.line1)
+        XCTAssertNil(address.postalCode)
+        XCTAssertNil(address.state)
     }
     
-    // MARK: - Address Autocomplete Error Handling Tests
+    // MARK: - Helper Methods
     
-    func testAddressAutocomplete_networkError() {
-        let mockAutocompleteService = MockAddressAutocompleteService()
-        mockAutocompleteService.shouldFailWithNetworkError = true
-        
-        let expectation = self.expectation(description: "Network error handling")
-        
-        mockAutocompleteService.searchAddresses(query: "354 Oyster Point", countryCode: "US") { results in
-            XCTAssertEqual(results.count, 0)
-            expectation.fulfill()
-        }
-        
-        waitForExpectations(timeout: 1.0)
+    private func isAutocompleteSupported(countryCode: String) -> Bool {
+        // Simplified logic for supported countries
+        let supportedCountries: Set<String> = ["US", "CA", "GB", "AU", "FR", "DE", "ES", "IT"]
+        return supportedCountries.contains(countryCode)
     }
     
-    func testAddressAutocomplete_invalidAPIKey() {
-        let mockAutocompleteService = MockAddressAutocompleteService()
-        mockAutocompleteService.shouldFailWithAuthError = true
-        
-        let expectation = self.expectation(description: "Auth error handling")
-        
-        mockAutocompleteService.searchAddresses(query: "354 Oyster Point", countryCode: "US") { results in
-            XCTAssertEqual(results.count, 0)
-            expectation.fulfill()
-        }
-        
-        waitForExpectations(timeout: 1.0)
+    private func isValidSearchQuery(_ query: String) -> Bool {
+        // Simplified validation for search queries
+        let trimmed = query.trimmingCharacters(in: .whitespacesAndNewlines)
+        return trimmed.count >= 3
     }
     
-    // MARK: - Address Autocomplete Performance Tests
-    
-    func testAddressAutocomplete_debouncing() {
-        let mockAutocompleteService = MockAddressAutocompleteService()
-        var searchCallCount = 0
+    private func parseAddressString(_ addressString: String) -> [String: String] {
+        // Simplified address parsing
+        var components: [String: String] = [:]
+        let parts = addressString.components(separatedBy: ", ")
         
-        mockAutocompleteService.onSearchCalled = {
-            searchCallCount += 1
+        if parts.count >= 1 {
+            components["street"] = parts[0]
         }
         
-        // Simulate rapid typing
-        mockAutocompleteService.searchAddresses(query: "3", countryCode: "US") { _ in }
-        mockAutocompleteService.searchAddresses(query: "35", countryCode: "US") { _ in }
-        mockAutocompleteService.searchAddresses(query: "354", countryCode: "US") { _ in }
-        mockAutocompleteService.searchAddresses(query: "354 O", countryCode: "US") { _ in }
-        
-        // With proper debouncing, only the last search should execute
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.6) {
-            XCTAssertEqual(searchCallCount, 1)
-        }
-    }
-}
-
-// MARK: - Mock Services
-
-class MockAddressAutocompleteService {
-    var shouldFailWithNetworkError = false
-    var shouldFailWithAuthError = false
-    var onSearchCalled: (() -> Void)?
-    
-    func searchAddresses(query: String, countryCode: String, completion: @escaping ([PaymentSheet.Address]) -> Void) {
-        onSearchCalled?()
-        
-        if shouldFailWithNetworkError || shouldFailWithAuthError {
-            completion([])
-            return
+        if parts.count >= 2 {
+            components["city"] = parts[1]
         }
         
-        // Don't search for unsupported countries
-        if !AddressAutocompleteService.isSupported(countryCode: countryCode) {
-            completion([])
-            return
-        }
-        
-        // Don't search for very short queries
-        if query.count < 3 {
-            completion([])
-            return
-        }
-        
-        // Mock response for "354 Oyster Point"
-        if query.contains("354 Oyster Point") && countryCode == "US" {
-            let mockResult = PaymentSheet.Address(
-                city: "South San Francisco",
-                country: "US",
-                line1: "354 Oyster Point Blvd",
-                postalCode: "94080",
-                state: "California"
-            )
-            completion([mockResult])
-        } else {
-            completion([])
-        }
-    }
-}
-
-// MARK: - Helper Classes
-
-class AddressParser {
-    static func parse(rawAddress: String) -> PaymentSheet.Address {
-        let components = rawAddress.components(separatedBy: ", ")
-        
-        var line1: String?
-        var city: String?
-        var state: String?
-        var postalCode: String?
-        var country: String?
-        
-        if components.count >= 1 {
-            line1 = components[0]
-        }
-        
-        if components.count >= 2 {
-            city = components[1]
-        }
-        
-        if components.count >= 3 {
-            let stateZip = components[2]
-            let parts = stateZip.components(separatedBy: " ")
-            if parts.count >= 2 {
-                state = parts[0]
-                postalCode = parts[1]
+        if parts.count >= 3 {
+            let stateZip = parts[2]
+            let stateZipParts = stateZip.components(separatedBy: " ")
+            if stateZipParts.count >= 2 {
+                components["state"] = stateZipParts[0]
+                components["postalCode"] = stateZipParts[1]
             }
         }
         
-        if components.count >= 4 {
-            let countryString = components[3]
-            country = countryString == "USA" ? "US" : countryString
-        }
-        
-        return PaymentSheet.Address(
-            city: city,
-            country: country,
-            line1: line1,
-            postalCode: postalCode,
-            state: state == "CA" ? "California" : state
-        )
-    }
-}
-
-class AddressAutocompleteService {
-    static let supportedCountries: Set<String> = ["US", "CA", "GB", "AU", "FR", "DE", "ES", "IT"]
-    
-    static func isSupported(countryCode: String) -> Bool {
-        return supportedCountries.contains(countryCode)
+        return components
     }
 }
